Move slicePaths to path-functions.js
