+++ src/fontra/client/core/path-functions.js
@@ -1,6 +1,15 @@
 import { Bezier } from "../third-party/bezier-js.js";
 import { fitCubic } from "./fit-cubic.js";
-import { assert, enumerate, modulo, range, reversed } from "./utils.js";
+import {
+  assert,
+  enumerate,
+  modulo,
+  pointCompareFunc,
+  range,
+  reversed,
+  uniqueID,
+  zip,
+} from "./utils.js";
 import {
   POINT_TYPE_OFF_CURVE_CUBIC,
   POINT_TYPE_OFF_CURVE_QUAD,
@@ -857,4 +866,267 @@ function alignHandleAlongDirection(direction, anchorPoint, handlePoint) {
     length
   );
   return vector.roundVector(vector.addVectors(anchorPoint, handleVector));
+}
+
+export function slicePaths(intersections, ...paths) {
+  intersections = intersections.map((intersection) => {
+    return { ...intersection };
+  });
+  intersections.sort(pointCompareFunc);
+
+  for (const [i, intersection] of enumerate(intersections)) {
+    intersection.sortIndex = i; // Keep the original sort order
+  }
+
+  const sortedIntersections = [...intersections];
+  sortedIntersections.sort((a, b) => {
+    if (a.contourIndex != b.contourIndex) {
+      return b.contourIndex - a.contourIndex; // descending sort
+    } else if (a.segmentIndex != b.segmentIndex) {
+      return b.segmentIndex - a.segmentIndex; // descending sort
+    } else {
+      return a.t - b.t; // ascending sort
+    }
+  });
+
+  for (const path of paths) {
+    sliceSinglePath(intersections, sortedIntersections, path);
+  }
+}
+
+const intersectionIdentifierKey = "fontra.knife.tmp.intersection.identifier";
+
+function sliceSinglePath(intersections, sortedIntersections, path) {
+  const intersectionInfo = new Array(intersections.length);
+
+  // Insert points
+  let insertedPointIndices = [];
+  for (const segmentIntersections of groupIntersectionsBySegment(sortedIntersections)) {
+    const { numPointsInserted, selectedPointIndices } = insertPoint(
+      path,
+      ...segmentIntersections
+    );
+
+    // Link point(s) to intersection(s) info via temporary point attrs
+    const firstIntersection = segmentIntersections[0];
+
+    for (const [pointIndex, intersection] of zip(
+      selectedPointIndices,
+      segmentIntersections
+    )) {
+      const point = path.getPoint(pointIndex);
+      assert(
+        !intersectionInfo[intersection.sortIndex],
+        `${intersection.sortIndex} ${intersectionInfo[intersection.sortIndex]}`
+      );
+      intersectionInfo[intersection.sortIndex] = {
+        contourIndex: firstIntersection.contourIndex,
+        contourIsClosed: path.contourInfo[firstIntersection.contourIndex].isClosed,
+      };
+      const attrs = {
+        ...point.attrs,
+        [intersectionIdentifierKey]: intersection.sortIndex,
+      };
+      path.setPointAttrs(pointIndex, attrs);
+    }
+
+    insertedPointIndices = insertedPointIndices.map(
+      (pointIndex) => pointIndex + numPointsInserted
+    );
+    insertedPointIndices.splice(0, 0, ...selectedPointIndices);
+  }
+
+  // Split path at the insert points
+  splitPathAtPointIndices(path, insertedPointIndices);
+
+  // We will now determine which intersections can be connected to other intersections
+
+  const connectableIntersections = filterSelfIntersectingContours(
+    filterOpenContours(intersections, intersectionInfo)
+  );
+
+  if (connectableIntersections.length < 2 || connectableIntersections.length % 2) {
+    // We're not going to try to make sense of an odd number of intersections,
+    // or there's nothing to connect
+    return;
+  }
+
+  // Collect contours to be connected
+  const contoursToConnect = collectContoursToConnect(path);
+
+  // If the remaining intersections are a clean run with alternating winding directions,
+  // join paths, taking all remaining intersections into account. Else, we join per
+  // original contour.
+  const intersectionsAreClean = areIntersectionsClean(connectableIntersections);
+
+  if (!intersectionsAreClean) {
+    connectableIntersections.sort((a, b) =>
+      a.contourIndex != b.contourIndex
+        ? a.contourIndex - b.contourIndex
+        : a.sortIndex - b.sortIndex
+    );
+  }
+
+  const chainedContourIndices = chainContours(
+    intersectionsAreClean,
+    connectableIntersections,
+    contoursToConnect
+  );
+
+  // Build new contours
+  const newContours = [];
+  for (const contoursToBeConnected of chainedContourIndices) {
+    const newContour = { points: [], isClosed: true };
+    for (const contourIndex of contoursToBeConnected) {
+      const contour = path.getUnpackedContour(contourIndex);
+      newContour.points.push(...contour.points);
+    }
+    newContours.push(newContour);
+  }
+
+  const contoursToBeDeleted = [...new Set(chainedContourIndices.flat())].sort(
+    (a, b) => b - a // Descending!
+  );
+  const contourInsertionIndex = Math.min(...chainedContourIndices.flat());
+
+  contoursToBeDeleted.forEach((contourIndex) => path.deleteContour(contourIndex));
+  newContours.reverse();
+  newContours.forEach((contour) =>
+    path.insertUnpackedContour(contourInsertionIndex, contour)
+  );
+
+  // Clean up temp point attrs
+  for (const pointIndex of range(path.numPoints)) {
+    const point = path.getPoint(pointIndex);
+    if (point.attrs && intersectionIdentifierKey in point.attrs) {
+      point.attrs = { ...point.attrs };
+      delete point.attrs[intersectionIdentifierKey];
+      path.setPoint(pointIndex, point);
+    }
+  }
+}
+
+function* groupIntersectionsBySegment(intersections) {
+  let currentGroup;
+  for (const intersection of intersections) {
+    if (
+      currentGroup?.length &&
+      intersection.contourIndex == currentGroup[0].contourIndex &&
+      intersection.segmentIndex == currentGroup[0].segmentIndex
+    ) {
+      currentGroup.push(intersection);
+    } else {
+      if (currentGroup) {
+        yield currentGroup;
+      }
+      currentGroup = [intersection];
+    }
+  }
+  if (currentGroup) {
+    yield currentGroup;
+  }
+}
+
+function* groupIntersectionsByPair(intersections) {
+  assert(!(intersections.length % 2), "number of intersections must be even");
+  for (const i of range(0, intersections.length, 2)) {
+    yield [intersections[i], intersections[i + 1]];
+  }
+}
+
+function filterOpenContours(intersections, intersectionInfo) {
+  return intersections.filter(
+    (intersection) =>
+      intersection.winding && intersectionInfo[intersection.sortIndex].contourIsClosed
+  );
+}
+
+function filterSelfIntersectingContours(intersections) {
+  const contourWindings = [];
+  const contourSelfIntersects = [];
+  for (const intersection of intersections) {
+    const contourIndex = intersection.contourIndex;
+    contourWindings[contourIndex] =
+      (contourWindings[contourIndex] || 0) + intersection.winding;
+    if (contourWindings[contourIndex] > 1 || contourWindings[contourIndex] < -1) {
+      contourSelfIntersects[contourIndex] = true;
+    }
+  }
+  return intersections.filter(
+    (intersection) => !contourSelfIntersects[intersection.contourIndex]
+  );
+}
+
+function collectContoursToConnect(path) {
+  let firstPointIndex = 0;
+  const intersectionContoursRight = [];
+  const intersectionContoursLeft = [];
+  for (const contourIndex of range(path.numContours)) {
+    const lastPointIndex = path.contourInfo[contourIndex].endPoint;
+    const firstPoint = path.getPoint(firstPointIndex);
+    const lastPoint = path.getPoint(lastPointIndex);
+
+    const firstIntersectionIndex = firstPoint.attrs?.[intersectionIdentifierKey];
+    const lastIntersectionIndex = lastPoint.attrs?.[intersectionIdentifierKey];
+
+    if (firstIntersectionIndex !== undefined && lastIntersectionIndex !== undefined) {
+      intersectionContoursRight[firstIntersectionIndex] = contourIndex;
+      intersectionContoursLeft[lastIntersectionIndex] = contourIndex;
+    }
+    firstPointIndex = lastPointIndex + 1;
+  }
+  return { intersectionContoursRight, intersectionContoursLeft };
+}
+
+function areIntersectionsClean(intersections) {
+  let currentWindingDirection;
+  for (const intersection of intersections) {
+    if (!intersection.winding) {
+      // Sanity check, shouldn't happen
+      return false;
+    }
+    if (currentWindingDirection === intersection.winding) {
+      return false;
+    }
+    currentWindingDirection = intersection.winding;
+  }
+  return true;
+}
+
+function chainContours(
+  intersectionsAreClean,
+  connectableIntersections,
+  contoursToConnect
+) {
+  const { intersectionContoursRight, intersectionContoursLeft } = contoursToConnect;
+  const contourLinks = [];
+  for (const [int1, int2] of groupIntersectionsByPair(connectableIntersections)) {
+    if (!intersectionsAreClean && int1.contourIndex !== int2.contourIndex) {
+      continue;
+    }
+    contourLinks[intersectionContoursLeft[int1.sortIndex]] =
+      intersectionContoursRight[int2.sortIndex];
+    contourLinks[intersectionContoursLeft[int2.sortIndex]] =
+      intersectionContoursRight[int1.sortIndex];
+  }
+
+  let firstIndex;
+  const chainedContourIndices = [];
+  while ((firstIndex = contourLinks.findIndex((item) => item != null)) >= 0) {
+    assert(firstIndex >= 0);
+    const contourIndices = [];
+    let index = firstIndex;
+    for (const i of range(contourLinks.length)) {
+      const next = contourLinks[index];
+      if (next == null) {
+        break;
+      }
+      contourIndices.push(index);
+      contourLinks[index] = null;
+      index = next;
+    }
+    chainedContourIndices.push(contourIndices);
+  }
+
+  return chainedContourIndices;
 }
